Adding in gravity-affected physics objects and player death on hazard collision
