+++ UpsideDownGJ/Assets/Prefabs/GameController.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a3d828db74def4cc2b457cb5e0d0ea20
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
