+++ UpsideDownGJ/Assets/Prefabs/Spike.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e2eed5d3078624d93b23aad13d8c7497
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
