+++ UpsideDownGJ/ProjectSettings/EditorBuildSettings.asset
@@ -8,4 +8,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/SampleScene.unity
     guid: 2cda990e2423bbf4892e6590ba056729
+  - enabled: 1
+    path: Assets/Scenes/gravity_switch.unity
+    guid: 8cda964b8690040eaba2176d1bd270ad
   m_configObjects: {}
